+++ libdariadb/engine.h
@@ -5,6 +5,7 @@
 #include <libdariadb/storage/settings.h>
 #include <libdariadb/utils/utils.h>
 #include <libdariadb/st_exports.h>
+#include <libdariadb/storage/strategy.h>
 #include <memory>
 
 namespace dariadb {
@@ -63,8 +64,13 @@ public:
 
   EXPORT void fsck();
 
-  EXPORT Version version();
   EXPORT void eraseOld(const Time&t);
+
+  EXPORT void compactTo(uint32_t pagesCount);
+  EXPORT void compactbyTime(Time from, Time to);
+
+  EXPORT Version version();
+  EXPORT STRATEGY strategy()const;
 protected:
   class Private;
   std::unique_ptr<Private> _impl;