+++ libdariadb/interfaces/ichunkcontainer.h
@@ -9,7 +9,8 @@ namespace storage {
 
 struct ChunkLink {
   uint64_t id;
-  uint64_t id_bloom;
+  uint64_t meas_id;
+  dariadb::Time minTime;
   dariadb::Time maxTime;
   std::string page_name;
   uint64_t index_rec_number;